implement (desktop_entry_tree_foreach): implement
2003-05-01  Havoc Pennington  <hp@pobox.com>

	* src/menu-process.c (desktop_entry_tree_print): implement
	(desktop_entry_tree_foreach): implement

	* src/menu-entries.c (entry_set_new): fix to init to all bits zero

+++ src/gen-compat-tree.c
@@ -284,7 +284,7 @@ markup_get_root_node_name (const char  *filename,
   FILE *f;
 #define BUFLEN 128
   char buf[BUFLEN];
-  int bytes_read;
+  size_t bytes_read;
   
   root_element = NULL;
